ml-service/requirements.txt
# ML Service Requirements for Smart Crop Advisory Pest Detection
# Core ML Libraries
tensorflow>=2.13.0
keras>=2.13.1
numpy>=1.24.0
opencv-python>=4.8.0
pillow>=10.0.0
scikit-learn>=1.3.0

# Web Framework
flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0

# Image Processing
opencv-contrib-python>=4.8.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Data Processing
pandas>=2.0.0
scipy>=1.11.0

# Model Deployment (tf-serving-api removed - not needed for basic deployment)

# Utilities
python-dotenv>=1.0.0
gunicorn>=21.0.0